rename NavBar, MenuBar, hotkeys
