+++ client/src/common/components/Embedding/components/Opacities/components/Opacity/Opacity.tsx
