+++ client/src/components/MenuBarTemp/components/Clip/Clip.tsx
@@ -11,7 +11,7 @@ import {
 } from "@blueprintjs/core";
 import { IconNames } from "@blueprintjs/icons";
 
-import { tooltipHoverOpenDelay } from "../../../globals";
+import { tooltipHoverOpenDelay } from "~/globals";
 // @ts-expect-error ts-migrate(2307) FIXME: Cannot find module './menubar.css' or its correspo... Remove this comment to see the full error message
 import styles from "../menubar.css";
 import { ClipProps } from "./types";
